requirements.txt
pyoai==2.5.0
requests==2.18.4
Jinja2==2.10
boto3==1.6.6
tqdm==4.19.6
python_dateutil==2.6.1
pre-commit=1.7.0
